+++ src/defaultSettings.ts
